Refactor to separate DOH<->UDP from VPN logic
This is a step toward enabling access to DOH from a UDP
server.
