+++ Android/app/src/main/java/app/intra/util/DnsUdpQuery.java
@@ -15,10 +15,17 @@ limitations under the License.
 */
 package app.intra.util;
 
+import com.google.firebase.crash.FirebaseCrash;
+
+import android.os.SystemClock;
+import android.util.Log;
+
 import java.net.InetAddress;
+import java.net.ProtocolException;
 
 // Class to maintain state about a DNS query over UDP.
 public class DnsUdpQuery {
+  private static final String LOG_TAG = "DnsUdpQuery";
 
   public String name;
   public short requestId;
@@ -28,4 +35,33 @@ public class DnsUdpQuery {
   public short sourcePort;
   public short destPort;
   public long timestamp;
+
+  // Returns the question name, type, and id  present in |dnsPacket|, as a DnsUdpQuery object.
+  // Assumes the DNS packet has been validated.
+  public static DnsUdpQuery fromUdpBody(byte[] data) {
+    DnsUdpQuery dnsUdpQuery = new DnsUdpQuery();
+    dnsUdpQuery.timestamp = SystemClock.elapsedRealtime();
+
+    DnsPacket dnsPacket;
+    try {
+      dnsPacket = new DnsPacket(data);
+    } catch (ProtocolException e) {
+      FirebaseCrash.logcat(Log.INFO, LOG_TAG, "Received invalid DNS request");
+      return null;
+    }
+    if (!dnsPacket.isNormalQuery() && !dnsPacket.isResponse()) {
+      FirebaseCrash.logcat(Log.INFO, LOG_TAG, "Dropping strange DNS query");
+      return null;
+    }
+
+    dnsUdpQuery.type = dnsPacket.getQueryType();
+    dnsUdpQuery.name = dnsPacket.getQueryName();
+    if (dnsUdpQuery.name == null || dnsUdpQuery.type == 0) {
+      FirebaseCrash.logcat(Log.INFO, LOG_TAG, "No question in DNS packet");
+      return null;
+    }
+    dnsUdpQuery.requestId = dnsPacket.getId();
+
+    return dnsUdpQuery;
+  }
 }
